fixup! fixup! [v3-1-test] Add bundle_path temporarily to sys.path during DagBag parsing. (#55894)
